+++ CHANGELOG.md
@@ -9,7 +9,9 @@
 - Added retrieving query complexity once query has been completed (#316) 
 - Allow input types to be passed in from variables using \stdClass instead of associative arrays (#535)
 - Fixes parsing of string literals of the form \u0000 for code points in the range [128, 255] inclusive
-- Parse UTF-16 surrogate pairs within string literals 
+- Parse UTF-16 surrogate pairs within string literals
+- Renamed `GraphQL\Error\Debug` to `GraphQL\Error\DebugFlag`. 
+- Debug flags in `GraphQL\Executor\ExecutionResult`, `GraphQL\Error\FormattedError` and `GraphQL\Server\ServerConfig` do not accept `boolean` value anymore but `int` only. 
 
 #### v0.13.5
 - Fix coroutine executor when using with promise (#486)